backend/requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
pydantic>=2.11.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
crewai>=0.130.0
litellm>=1.74.0
pymongo[srv]>=4.10.1,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
email-validator>=2.2.0,<3.0.0
httpx>=0.28.1,<1.0.0
